requirements.txt
Flask==2.3.2
flask-sqlalchemy==3.0.3
pyodbc==4.0.39
flask-jwt-extended==4.5.2
xmltodict==0.13.0
pandas==2.0.1
openpyxl==3.1.2
python-dotenv==1.0.0
marshmallow==3.19.1